requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.10.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
databases==0.9.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
